# docs/swagger.yaml
basePath: /
definitions:
  auth.GetTokenDto:
    properties:
      email:
        example: ingshelpidy@gmail.com
        type: string
    required:
    - email
    type: object
  auth.TokenResponse:
    properties:
      message:
        type: string
      token:
        type: string
      user: {}
    type: object
...
  models.ErrorResponse:
    description: Error response structure with a message.
    properties:
      message:
        type: string
    type: object
...
host: localhost:5005
info:
  contact: {}
  description: This is the API documentation for the Phinex Blog Application.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Phinex Blog API
  version: "1.0"
paths:
  /auth/token:
    post:
      consumes:
      - application/json
      description: Generates a JWT token for a verified user by email.
      parameters:
      - description: User email for token generation
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/auth.GetTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get JWT token by email
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"